+++ source/unit_mod_web_util.pas
@@ -12,8 +12,8 @@ implementation
 uses
   Variants, unit_js, unit_thread,
   uCEFTypes, uCEFInterfaces, unit_global, LazFileUtils,
-  uCEFConstants, uCEFv8Value, uCEFValue, uCefDictionaryValue, uCefBinaryValue,
-  uCEFUrlRequestClientComponent, uCEFRequest, uCEFUrlRequest;
+  uCEFConstants, uCEFv8Value, uCEFValue, uCefDictionaryValue, uCefListValue,
+  uCEFUrlRequestClientComponent, uCEFRequest, uCEFUrlRequest, uCefChromium;
 
 const
   MODULE_NAME = 'web_util'; //////////
@@ -28,6 +28,10 @@ begin
   handler:= TV8HandlerSafe.Create(UTF8Encode(name), 'downloadFile');
   Result.SetValueByKey('downloadFile',
    TCefv8ValueRef.NewFunction('downloadFile', handler), V8_PROPERTY_ATTRIBUTE_NONE);
+
+  handler:= TV8HandlerSafe.Create(UTF8Encode(name), 'scraping');
+  Result.SetValueByKey('scraping',
+   TCefv8ValueRef.NewFunction('scraping', handler), V8_PROPERTY_ATTRIBUTE_NONE);
 end;
 
 type
@@ -60,6 +64,62 @@ type
     procedure ExecuteAct; override;
   end;
 
+  { TScrapingThread }
+
+  TScrapingThread = class(TPromiseThread)
+  protected
+    procedure ExecuteAct; override;
+  public
+  end;
+
+  { TScrapingGetSourceThread }
+
+  TScrapingGetSourceThread = class(TPromiseThread)
+  protected
+    procedure ExecuteAct; override;
+  public
+  end;
+
+  { TScrapingNewFunctionThread }
+
+  TScrapingNewFunctionThread = class(TPromiseThread)
+  protected
+    procedure ExecuteAct; override;
+  public
+  end;
+
+  { TScrapingWaitThread }
+
+  TScrapingWaitThread = class(TPromiseThread)
+  protected
+    procedure ExecuteAct; override;
+  public
+  end;
+
+  { TScrapingPrepareReloadThread }
+
+  TScrapingPrepareReloadThread = class(TPromiseThread)
+  protected
+    procedure ExecuteAct; override;
+  public
+  end;
+
+  { TScrapingCancelThread }
+
+  TScrapingCancelThread = class(TPromiseThread)
+  protected
+    procedure ExecuteAct; override;
+  public
+  end;
+
+  { TScrapingSendKeyPressThread }
+
+  TScrapingSendKeyPressThread = class(TPromiseThread)
+  protected
+    procedure ExecuteAct; override;
+  public
+  end;
+
 //
 function safeExecute(const handler: TV8HandlerSafe; const name: ustring;
   const obj: ICefv8Value; const arguments: TCefv8ValueArray;
@@ -69,6 +129,13 @@ begin
   try
     case handler.FuncName of
 
+     'scraping',
+     'scraping.getSource',
+     'scraping.newFunction',
+     'scraping.wait',
+     'scraping.prepareReload',
+     'scraping.sendKeyPress',
+     'scraping.cancel',
      'request.read',
      'request.cancel',
      'downloadFile': begin
@@ -103,11 +170,34 @@ begin
     'request.cancel': begin
       StartPromiseThread(TCancelThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
     end;
+    'scraping': begin
+      StartPromiseThread(TScrapingThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
+    end;
+    'scraping.getSource': begin
+      StartPromiseThread(TScrapingGetSourceThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
+    end;
+    'scraping.newFunction': begin
+      StartPromiseThread(TScrapingNewFunctionThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
+    end;
+    'scraping.wait': begin
+      StartPromiseThread(TScrapingWaitThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
+    end;
+    'scraping.prepareReload': begin
+      StartPromiseThread(TScrapingPrepareReloadThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
+    end;
+    'scraping.sendKeyPress': begin
+      StartPromiseThread(TScrapingSendKeyPressThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
+    end;
+    'scraping.cancel': begin
+      StartPromiseThread(TScrapingCancelThread, Args, arguments[0], arguments[1], ModuleName, FuncName, CefObject);
+    end;
     else
       Exit;
   end;
   Result:= True;
 end;
+
+// downloadFile ////////////////////////////////////////////////////////////////
 
 { TRequestObject }
 
@@ -236,13 +326,13 @@ begin
   dic:= NewUserObject(reqObj);
 
   func:= TCefDictionaryValueRef.New;
-  func.SetBool(VTYPE_FUNCTION_NAME, true);
+  func.SetString(VTYPE_FUNCTION_NAME, 'read');
   func.SetString('ModuleName', MODULE_NAME);
   func.SetString('FuncName', 'request.read');
   dic.SetDictionary('read', func);
 
   func:= TCefDictionaryValueRef.New;
-  func.SetBool(VTYPE_FUNCTION_NAME, true);
+  func.SetString(VTYPE_FUNCTION_NAME, 'cancel');
   func.SetString('ModuleName', MODULE_NAME);
   func.SetString('FuncName', 'request.cancel');
   dic.SetDictionary('cancel', func);
@@ -258,7 +348,6 @@ var
   obj: TObject;
   req: TRequestObject;
   option, dic: ICefDictionaryValue;
-  bin: ICefBinaryValue;
   wait, waitc: integer;
 begin
   dic:= CefObject.GetDictionary;
@@ -299,8 +388,6 @@ begin
   end;
 end;
 
-{ TCancelThread }
-
 procedure TCancelThread.ExecuteAct;
 var
   obj: TObject;
@@ -321,11 +408,383 @@ begin
   CefResolve.SetBool(true);
 end;
 
+// scraping ////////////////////////////////////////////////////////////////////
+
+type
+
+  { TScrapingObject }
+
+  TScrapingObject = class
+    crm: TChromium;
+    loaded, canceled, contextCreated, errorOccurred: boolean;
+    FSource, FErrorText: ustring;
+    destructor Destroy; override;
+    procedure ChromiumLoadStart(Sender: TObject; const browser: ICefBrowser; const frame: ICefFrame; transitionType: TCefTransitionType);
+    procedure ChromiumLoadEnd(Sender: TObject; const Browser: ICefBrowser;
+     const Frame: ICefFrame; httpStatusCode: Integer);
+    procedure ChromiumLoadError(Sender: TObject; const browser: ICefBrowser; const frame: ICefFrame; errorCode: TCefErrorCode; const errorText, failedUrl: ustring);
+    procedure ChromiumProcessMessageReceived(Sender: TObject; const browser: ICefBrowser; const frame: ICefFrame; sourceProcess: TCefProcessId; const message: ICefProcessMessage; out Result: Boolean);
+    procedure ChromiumTextResultAvailableEvent(Sender: TObject; const aText : ustring);
+    procedure ChromiumConsoleMessage(Sender: TObject; const browser: ICefBrowser; level: TCefLogSeverity; const message, source: ustring; line: Integer; out Result: Boolean);
+  end;
+
+{ TScrapingObject }
+
+destructor TScrapingObject.Destroy;
+begin
+  crm.Free;
+  inherited Destroy;
+end;
+
+procedure TScrapingObject.ChromiumLoadStart(Sender: TObject;
+  const browser: ICefBrowser; const frame: ICefFrame;
+  transitionType: TCefTransitionType);
+begin
+  loaded:= false;
+  errorOccurred:= false;
+end;
+
+procedure TScrapingObject.ChromiumLoadEnd(Sender: TObject;
+  const Browser: ICefBrowser; const Frame: ICefFrame; httpStatusCode: Integer);
+begin
+  if (Frame = nil) or not Frame.IsValid or not Frame.IsMain then Exit;
+  crm.RetrieveHTML();
+end;
+
+procedure TScrapingObject.ChromiumLoadError(Sender: TObject;
+  const browser: ICefBrowser; const frame: ICefFrame; errorCode: TCefErrorCode;
+  const errorText, failedUrl: ustring);
+begin
+  FErrorText:= 'Load Error ' +
+   UTF8Decode(IntToStr(errorCode)) + ': ' +
+   errorText + #$0d +
+   failedUrl;
+  errorOccurred:= true;
+end;
+
+procedure TScrapingObject.ChromiumProcessMessageReceived(Sender: TObject;
+  const browser: ICefBrowser; const frame: ICefFrame;
+  sourceProcess: TCefProcessId; const message: ICefProcessMessage; out
+  Result: Boolean);
+
+  //
+  procedure context_created;
+  //var
+  //  argv: string;
+  //  us: ustring;
+  //  i: integer;
+  begin
+    //argv:= '[';
+    //for i:= 0 to ParamCount do begin
+    //  argv:= argv + StringPasToJS2(ParamStr(i)) + ',';
+    //end;
+    //argv:= argv + ']';
+    //
+    //us:= UTF8Decode(''
+    // + '{'
+    // + 'const v=window.' + G_VAR_IN_JS_NAME + ';'
+    // + 'v._ipc={};'
+    // + 'v._ipc_g={};'
+    // + 'window.__dogroot = ' + StringPasToJS(dogroot) + ';'
+    // + 'window.__restroot = ' + StringPasToJS(restroot) + ';'
+    // + 'window.__execPath = ' + StringPasToJS(execPath) + ';'
+    // + 'window.__argv = ' + argv + ';'
+    // + '}'
+    //);
+    //Chromium.ExecuteJavaScript(us, 'about:blank');
+
+    contextCreated:= true;
+  end;
+
+  //
+  procedure new_function_re;
+  var
+    obj: TObjectWithInterface;
+  begin
+    obj:= TObjectWithInterface.Create;
+    obj.cefBaseRefCounted:= message.ArgumentList.GetValue(1);
+    SetObjectList(UTF8Encode(message.ArgumentList.GetString(0)), obj);
+  end;
+
+//
+begin
+  Result:= False;
+  case message.name of
+    'context_created': context_created;
+    'new_function_re': new_function_re;
+  end;
+  Result:= True;
+end;
+
+procedure TScrapingObject.ChromiumTextResultAvailableEvent(Sender: TObject;
+  const aText: ustring);
+begin
+  FSource:= aText;
+  loaded:= true;
+end;
+
+procedure TScrapingObject.ChromiumConsoleMessage(Sender: TObject;
+  const browser: ICefBrowser; level: TCefLogSeverity; const message,
+  source: ustring; line: Integer; out Result: Boolean);
+begin
+  if level = LOGSEVERITY_ERROR then begin
+    FErrorText:= message;
+    errorOccurred:= true;
+  end;
+end;
+
+{ TScrapingThread }
+
+procedure TScrapingThread.ExecuteAct;
+var
+  uobj: TScrapingObject;
+  option, dic, func: ICefDictionaryValue;
+begin
+  if Args.GetSize < 1 then Raise Exception.Create(ERROR_INVALID_PARAM_COUNT);
+
+  uobj:= TScrapingObject.Create;
+  uobj.crm:= TChromium.Create(nil);
+  uobj.crm.DefaultUrl:= Args.GetString(0);
+  uobj.crm.MultiBrowserMode:= false;
+  uobj.crm.WebRTCIPHandlingPolicy:= hpDisableNonProxiedUDP;
+  uobj.crm.WebRTCMultipleRoutes:= STATE_DISABLED;
+  uobj.crm.WebRTCNonproxiedUDP:= STATE_DISABLED;
+  uobj.crm.OnLoadStart:= @uobj.ChromiumLoadStart;
+  uobj.crm.OnLoadEnd:= @uobj.ChromiumLoadEnd;
+  uobj.crm.OnLoadError:= @uobj.ChromiumLoadError;
+  uobj.crm.OnTextResultAvailable:= @uobj.ChromiumTextResultAvailableEvent;
+  uobj.crm.OnProcessMessageReceived:= @uobj.ChromiumProcessMessageReceived;
+  uobj.crm.OnConsoleMessage:= @uobj.ChromiumConsoleMessage;
+
+  uobj.crm.CreateBrowser();
+
+  dic:= NewUserObject(uobj);
+
+  func:= TCefDictionaryValueRef.New;
+  func.SetString(VTYPE_FUNCTION_NAME, 'getSource');
+  func.SetString('ModuleName', MODULE_NAME);
+  func.SetString('FuncName', 'scraping.getSource');
+  dic.SetDictionary('getSource', func);
+
+  func:= TCefDictionaryValueRef.New;
+  func.SetString(VTYPE_FUNCTION_NAME, 'newFunction');
+  func.SetString('ModuleName', MODULE_NAME);
+  func.SetString('FuncName', 'scraping.newFunction');
+  dic.SetDictionary('newFunction', func);
+
+  func:= TCefDictionaryValueRef.New;
+  func.SetString(VTYPE_FUNCTION_NAME, 'wait');
+  func.SetString('ModuleName', MODULE_NAME);
+  func.SetString('FuncName', 'scraping.wait');
+  dic.SetDictionary('wait', func);
+
+  func:= TCefDictionaryValueRef.New;
+  func.SetString(VTYPE_FUNCTION_NAME, 'prepareReload');
+  func.SetString('ModuleName', MODULE_NAME);
+  func.SetString('FuncName', 'scraping.prepareReload');
+  dic.SetDictionary('prepareReload', func);
+
+  func:= TCefDictionaryValueRef.New;
+  func.SetString(VTYPE_FUNCTION_NAME, 'sendKeyPress');
+  func.SetString('ModuleName', MODULE_NAME);
+  func.SetString('FuncName', 'scraping.sendKeyPress');
+  dic.SetDictionary('sendKeyPress', func);
+
+  func:= TCefDictionaryValueRef.New;
+  func.SetString(VTYPE_FUNCTION_NAME, 'cancel');
+  func.SetString('ModuleName', MODULE_NAME);
+  func.SetString('FuncName', 'scraping.cancel');
+  dic.SetDictionary('cancel', func);
+
+  CefResolve:= TCefValueRef.New;
+  CefResolve.SetDictionary(dic);
+end;
+
+{ TScrapingGetSourceThread }
+
+procedure TScrapingGetSourceThread.ExecuteAct;
+var
+  obj: TObject;
+  uobj: TScrapingObject;
+  dic: ICefDictionaryValue;
+begin
+  dic:= CefObject.GetDictionary;
+  if not Assigned(dic) or not dic.IsValid then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+  obj:= GetObjectList(UTF8Encode(dic.GetString(VTYPE_OBJECT_NAME)));
+  if not Assigned(obj) or not(obj is TScrapingObject) then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+
+  uobj:= TScrapingObject(obj);
+
+  CefResolve:= TCefValueRef.New;
+  CefResolve.SetString(uobj.FSource);
+end;
+
+{ TScrapingNewFunctionThread }
+
+procedure TScrapingNewFunctionThread.ExecuteAct;
+var
+  obj: TObject;
+  uobj: TScrapingObject;
+  dic: ICefDictionaryValue;
+  list: ICefListValue;
+  i, len: integer;
+begin
+  if Args.GetSize < 1 then Raise Exception.Create(ERROR_INVALID_PARAM_COUNT);
+
+  dic:= CefObject.GetDictionary;
+  if not Assigned(dic) or not dic.IsValid then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+  obj:= GetObjectList(UTF8Encode(dic.GetString(VTYPE_OBJECT_NAME)));
+  if not Assigned(obj) or not(obj is TScrapingObject) then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+
+  uobj:= TScrapingObject(obj);
+
+  uobj.errorOccurred:= false;
+
+  list:= TCefListValueRef.New;
+  len:= Args.GetSize;
+  for i:= 1 to len-1 do list.SetValue(i-1, Args.GetValue(i));
+  CefResolve:= newFunctionRe(UTF8Encode(Args.GetString(0)), list, '', true, uobj.crm);
+end;
+
+{ TScrapingWaitThread }
+
+procedure TScrapingWaitThread.ExecuteAct;
+var
+  obj: TObject;
+  uobj: TScrapingObject;
+  dic: ICefDictionaryValue;
+  timeout, timeoutc: integer;
+begin
+  dic:= CefObject.GetDictionary;
+  if not Assigned(dic) or not dic.IsValid then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+  obj:= GetObjectList(UTF8Encode(dic.GetString(VTYPE_OBJECT_NAME)));
+  if not Assigned(obj) or not(obj is TScrapingObject) then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+
+  timeout:= 30;
+  if Args.GetSize >= 1 then begin
+    timeout:= Args.GetInt(0);
+    if timeout < 1 then timeout:= 0;
+    if timeout > 10000 then timeout:= 10000;
+  end;
+  timeout:= timeout * 100;
+
+  uobj:= TScrapingObject(obj);
+  timeoutc:= 0;
+  while true do begin
+    if Self.Terminated or unit_global.appClosing or
+      uobj.canceled or uobj.errorOccurred then break;
+    if uobj.contextCreated and uobj.loaded and not uobj.crm.IsLoading then break;
+    inc(timeoutc);
+    if timeoutc > timeout then break;
+    Sleep(10);
+  end;
+
+  dic:= TCefDictionaryValueRef.New;
+  //dic.SetBool('complete', true);
+  dic.SetBool('canceled', uobj.canceled);
+  //dic.SetBool('finished', true);
+  if uobj.errorOccurred then dic.SetString('error', uobj.FErrorText);
+  CefResolve:= TCefValueRef.New;
+  CefResolve.SetDictionary(dic);
+end;
+
+{ TScrapingPrepareReloadThread }
+
+procedure TScrapingPrepareReloadThread.ExecuteAct;
+var
+  obj: TObject;
+  uobj: TScrapingObject;
+  dic: ICefDictionaryValue;
+begin
+  dic:= CefObject.GetDictionary;
+  if not Assigned(dic) or not dic.IsValid then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+  obj:= GetObjectList(UTF8Encode(dic.GetString(VTYPE_OBJECT_NAME)));
+  if not Assigned(obj) or not(obj is TScrapingObject) then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+
+  uobj:= TScrapingObject(obj);
+  uobj.loaded:= false;
+  uobj.contextCreated:= false;
+
+  CefResolve:= TCefValueRef.New;
+  CefResolve.SetBool(true);
+end;
+
+{ TScrapingCancelThread }
+
+procedure TScrapingCancelThread.ExecuteAct;
+var
+  obj: TObject;
+  uobj: TScrapingObject;
+  dic: ICefDictionaryValue;
+begin
+  dic:= CefObject.GetDictionary;
+  if not Assigned(dic) or not dic.IsValid then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+  obj:= GetObjectList(UTF8Encode(dic.GetString(VTYPE_OBJECT_NAME)));
+  if not Assigned(obj) or not(obj is TScrapingObject) then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+
+  uobj:= TScrapingObject(obj);
+  uobj.canceled:= true;
+
+  CefResolve:= TCefValueRef.New;
+  CefResolve.SetBool(true);
+end;
+
+{ TScrapingSendKeyPressThread }
+
+procedure TScrapingSendKeyPressThread.ExecuteAct;
+var
+  obj: TObject;
+  uobj: TScrapingObject;
+  dic: ICefDictionaryValue;
+  key, modifiers: integer;
+  kev: TCefKeyEvent;
+begin
+  if Args.GetSize < 1 then Raise Exception.Create(ERROR_INVALID_PARAM_COUNT);
+
+  dic:= CefObject.GetDictionary;
+  if not Assigned(dic) or not dic.IsValid then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+  obj:= GetObjectList(UTF8Encode(dic.GetString(VTYPE_OBJECT_NAME)));
+  if not Assigned(obj) or not(obj is TScrapingObject) then
+    Raise Exception.Create(ERROR_INVALID_HANDLE_VALUE);
+
+  uobj:= TScrapingObject(obj);
+
+  key:= Args.GetInt(0);
+  modifiers:= 0;
+  if Args.GetSize >= 2 then modifiers:= Args.GetInt(1);
+
+  FillChar(kev{%H-}, SizeOf(kev), 0);
+  kev.windows_key_code:= key;
+  kev.native_key_code:= key;
+  //kev.character:= Chr(key);
+  //kev.unmodified_character:= Chr(key);
+  kev.modifiers:= modifiers;
+  kev.focus_on_editable_field:= 0;
+  kev.is_system_key:= 0;
+  kev.kind:= KEYEVENT_CHAR;
+  uobj.crm.SendKeyEvent(@kev);
+
+  CefResolve:= TCefValueRef.New;
+  CefResolve.SetBool(true);
+end;
+
 //
 const
   _import = G_VAR_IN_JS_NAME + '["' + MODULE_NAME + '"]';
   _body = _import + '.__init__();' +
      'export const downloadFile=' + _import + '.downloadFile;' +
+     'export const scraping=' + _import + '.scraping;' +
      '';
 
 initialization
@@ -336,5 +795,12 @@ initialization
   AddPromiseThreadClass(MODULE_NAME, TDownloadFileThread);
   AddPromiseThreadClass(MODULE_NAME, TReadThread);
   AddPromiseThreadClass(MODULE_NAME, TCancelThread);
+  AddPromiseThreadClass(MODULE_NAME, TScrapingThread);
+  AddPromiseThreadClass(MODULE_NAME, TScrapingGetSourceThread);
+  AddPromiseThreadClass(MODULE_NAME, TScrapingNewFunctionThread);
+  AddPromiseThreadClass(MODULE_NAME, TScrapingWaitThread);
+  AddPromiseThreadClass(MODULE_NAME, TScrapingPrepareReloadThread);
+  AddPromiseThreadClass(MODULE_NAME, TScrapingCancelThread);
+  AddPromiseThreadClass(MODULE_NAME, TScrapingSendKeyPressThread);
 end.